+++ changelog.md
@@ -9,6 +9,11 @@ All notable changes to HVCWISE_TEA will be documented in this file.
 ### Removed
 ### Fixed
 
+## [add_tests] - 2024-11-25
+### Changed
+- Replace "" by "_" in macro-scenario
+- Update main.jl
+
 ## [add_tests] - 2024-11-22
 ### Changed
 - No temporary file written in the matlab_tools folder during the process